% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformations.R
\name{Logistic_like}
\alias{Logistic_like}
\title{Logistic-like Transformation}
\usage{
Logistic_like(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A transformed numeric vector.
}
\description{
Applies a logistic-like transformation (1 - exp(-a*x)).
}
