% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformations.R
\name{Min_score_normalize}
\alias{Min_score_normalize}
\title{Min-Score Normalization}
\usage{
Min_score_normalize(data)
}
\arguments{
\item{data}{A numeric matrix or data frame.}
}
\value{
A normalized matrix.
}
\description{
Normalizes data by subtracting the minimum value and dividing by the standard deviation for each row.
}
