% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformations.R
\name{Quantile_normalize}
\alias{Quantile_normalize}
\title{Quantile Normalization}
\usage{
Quantile_normalize(mat)
}
\arguments{
\item{mat}{A numeric matrix.}
}
\value{
A quantile normalized matrix.
}
\description{
Performs quantile normalization on a matrix.
Requires the 'preprocessCore' package.
}
