% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformations.R
\name{Ratio_shift}
\alias{Ratio_shift}
\title{Ratio Shift Transformation}
\usage{
Ratio_shift(data)
}
\arguments{
\item{data}{A numeric matrix or data frame.}
}
\value{
A shifted matrix.
}
\description{
Shifts values by subtracting the minimum value for each row.
}
