% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformations.R
\name{Tanh_scaled}
\alias{Tanh_scaled}
\title{Scaled Tanh Transformation}
\usage{
Tanh_scaled(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A transformed numeric vector.
}
\description{
Applies a scaled hyperbolic tangent transformation ((tanh(x) + 1) / 2).
}
